requirements.txt
pandas==2.3.0
python-dotenv==1.1.1
langchain-community==0.3.26
langchain-core==0.3.66
qdrant-client
sentence-transformers
minio==7.2.15
pypdfium2
pypdf==5.6.1
docx2txt
numpy==1.26.4
scikit-learn==1.7.0
streamlit==1.45.1
fastapi==0.115.14
langchain==0.3.26
pydantic==2.11.7
starlette==0.46.0
langchain-qdrant
